Move constants file out
